Refactored Unittests
